+++ src/brightsidebudget/account.py
@@ -98,11 +98,19 @@ class QName():
 class Account():
     """
     An Account represents a single financial entity where transactions occur.
-    It is basically a QName with optional tags.
+    It is basically a QName with optional tags and information.
+
+    short_qname is the shortest version of the qname that can be used in reports.
+    For example, if the qname is "Assets:Savings:Capital gain" and the short_qname
+    is "Savings:Capital gain", then the account will never be printed as "Capital gain"
+    in reports even if this name is unambiguous.
     """
     def __init__(self, *, qname: Union[QName, str],
+                 short_qname: Union[QName, str, None] = None,
                  tags: Union[dict[str, str], None] = None):
-        self._qname = qname if isinstance(qname, QName) else QName(qname)
+        self._qname = None
+        self._short_qname = None
+        self.update_qname(qname, short_qname)
         self._tags = tags or {}
 
     @property
@@ -112,12 +120,31 @@ class Account():
         """
         return self._qname
 
-    @qname.setter
-    def qname(self, value: Union[QName, str]):
-        if isinstance(value, QName):
-            self._qname = value
+    @property
+    def short_qname(self) -> QName:
+        """
+        The shortest qualified name of the account to be used in reports.
+        """
+        return self._short_qname
+
+    def update_qname(self, qname: Union[QName, str],
+                     short_qname: Union[QName, str, None] = None):
+        self._qname = qname if isinstance(qname, QName) else QName(qname)
+        if short_qname is None:
+            self._short_qname = QName(self._qname.qlist[-1])
+        elif isinstance(short_qname, str):
+            self._short_qname = QName(short_qname)
         else:
-            self._qname = QName(value)
+            self._short_qname = short_qname
+
+        s = self._short_qname
+        if s.depth > self._qname.depth:
+            raise ValueError("Short qname must be shorter or equal to the qname.")
+        if self._qname.qlist[-s.depth:] != s.qlist:
+            raise ValueError("Short qname must be a suffix of the qname.")
+
+    def update_short_qname(self, short_qname: Union[QName, str]):
+        self.update_qname(self._qname, short_qname)
 
     @property
     def tags(self) -> dict[str, str]: