+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html) with
 the exception that the versions 0.*.* may have breaking changes in minor versions.
 
+## [0.4.6]
+### Added
+- Added `short_qname` attribute to `Account` to limit the shortest qualified
+  name of an account. This is a breaking change since the `qname` attribute now
+  has to be modified with `update_qname` method.
+
 ## [0.4.5]
 ### Fixed
 - Fixed `adjust_for_bassertion` when called with a child parameter with a